+++ dico_toolbox/aims_tools.py
@@ -300,138 +300,6 @@ def volume_to_mesh(volume, smoothingFactor=2.0, aimsThreshold='96%',
     return mesh
 
 
-# def volume_to_mesh_experimental(
-#         volume,
-#         gaussian_blur_FWWM=0,
-#         threshold_absolute=None,
-#         threshold_quantile=0,
-#         translation=(0, 0, 0),
-#         decimation_params=dict(
-#             reduction_rate=99,
-#             max_clearance=3,
-#             max_error=1,
-#             feature_angle=180),
-#         smoothing_params=dict(
-#             algorithm='lowpass',
-#             iterations=30,
-#             rate=0.4,
-#             # NOT IMPLEMENTED
-#             # smoothing feature angle (in degrees) below which the vertex is not moved,
-#             # only for the Laplacian algorithm, between 0 and 180 degree [default=0]
-#             # laplacian_angle=0,
-#             # smoothing restoring force for the Simple Spring and Polygon Spring
-#             # algorithm, between 0 and 1 [default=0.2]
-#             # polygonspring_force=0.2
-#         )):
-#     """Create a mesh from a volume.
-
-#     Before calculating the mesh, the volume is first blurred with a gaussian filter and then thresholded.
-
-#     Args:
-#         volume (np.ndarray or aims.Volume): 3D binary image (if not binary, all vxels != 0 are set to 1)
-#         gaussian_blur_FWWM (numeric) : the gaussian blur filter's full width at half maximum.
-#         threshold_absolute all voxels lower than this value will be set to zero.
-#         threshold_quantile (float in [0,1]): voxels with intensity below this quantile will be set to zero.
-#             Applied only if threshold_absolute is None.
-#         decimation_params (dict, optional): decimation parameters.
-#             reduction_rate : expected % decimation reduction rate
-#             max_clearance : maximum clearance of the decimation
-#             max_error : maximum error distance from the original mesh (mm)
-#             feature_angle : feature angle (degrees), between 0 and 180
-#         smoothing_params (dict, optional): smoothing parameters
-#             type : smoothing alorithm's type (laplacian, simplespring, polygonspring or lowpass)
-#             iterations : smoothing number of iterations
-#             rate : smoothing moving rate at each iteration
-
-#     Returns:
-#         aims Mesh: A mesh obtained from the input volume
-#     """
-
-#     log.debug("Volume to mesh:")
-#     volume = volume_to_ndarray(volume).astype(float)
-#     shape = volume.shape
-
-#     # gaussiam blur
-#     log.debug(" - Gaussian blur")
-#     if gaussian_blur_FWWM > 0:
-#         sigma = gaussian_blur_FWWM/2.3548200450309493
-#         volume = gaussian_filter(volume, sigma)
-
-#     # normalization
-#     log.debug(" - Normalization")
-#     assert(volume.max()-volume.min() != 0)
-#     def normalize(x): return (x-x.min())/(x.max()-x.min())
-#     volume = normalize(volume)
-
-#     # threshold
-#     log.debug(" - threshold")
-#     if threshold_absolute is not None:
-#         q = threshold_absolute
-#     elif threshold_quantile > 0:
-#         q = np.quantile(volume[volume > 0], threshold_quantile)
-#     else:
-#         q = 0
-#     volume = (volume > q).astype(int)
-
-#     log.debug(" - add -1 border")
-#     # add a -1 pad required for successive steps
-#     # create an aims volume to use the fillBorder function
-#     aims_volume = aims.Volume_S16(*shape, 1, 1)
-#     aims_volume[:, :, :, 0] = (265*volume).astype(np.int16)  # copy data
-#     aims_volume.fillBorder(-1)
-
-#     m = aimsalgo.Mesher()
-
-#     # Ça renvoie un dict (label: liste de maillages, un par interface).
-#     mesher_output = m.doit(aims_volume)
-#     # WARGNING: the following assumes that mesher_output contains only one interface.
-#     # Merge the meshes
-#     meshes = list(mesher_output.values())[0]
-#     merged_mesh = meshes[0]
-#     for mesh in meshes[1:]:
-#         aims.SurfaceManip.meshMerge(merged_mesh, mesh)
-
-#     mesh = merged_mesh
-
-#     if mesh.size() == 0:
-#         raise Exception("The mesh is empty, try different parameters")
-
-#     # DECIMATE
-#     if decimation_params is not None:
-#         d = decimation_params
-#         print(d)
-#         m.setDecimation(d['reduction_rate'], d['max_clearance'],
-#                         d['max_error'], d['feature_angle'])
-
-#         m.decimate(mesh)
-
-#     # SMOOTH
-#     if smoothing_params is not None:
-#         smoothing_types = {
-#             "lowpass": m.LOWPASS,
-#             'laplacian': m.LAPLACIAN,
-#             "simplespring": m.SIMPLESPRING,
-#             "polygonspring": m.POLYGONSPRING
-#         }
-#         assert smoothing_params['algorithm'].lower() in smoothing_types,\
-#             "smoothing_param must be one of {}, got '{}'".format(
-#             smoothing_types.keys(), smoothing_params['algorithm'])
-
-#         d = smoothing_params
-#         m.setSmoothing(
-#             smoothing_types[d['algorithm']], d['iterations'], d['rate'])
-
-#         m.smooth(mesh)
-
-#     if translation != (0, 0, 0):
-#         mesh = PyMesh(mesh)
-#         # add translation
-#         mesh.vertices = mesh.vertices + translation
-#         mesh = mesh.to_aims_mesh()
-
-#     return mesh
-
-
 def bucket_to_mesh(bucket, smoothingFactor=0, aimsThreshold=1,
                    deciMaxError=0.5, deciMaxClearance=1.0, smoothIt=20, translation=(0, 0, 0), scale=30):
     """Generate the mesh of the input bucket.
@@ -505,63 +373,6 @@ def bucket_to_aligned_mesh(raw_bucket, talairach_dxyz, talairach_rot, talairach_
     aims.SurfaceManip.meshTransform(mesh, M2)
 
     return mesh
-
-# def bucket_to_mesh_experimental(
-#         bucket,
-#         gaussian_blur_FWWM=0,
-#         threshold_absolute=None,
-#         threshold_quantile=0,
-#         decimation_params=dict(
-#             reduction_rate=99,
-#             max_clearance=3,
-#             max_error=1,
-#             feature_angle=180),
-#         smoothing_params=dict(
-#             algorithm='lowpass',
-#             iterations=30,
-#             rate=0.4,
-#             # NOT IMPLEMENTED
-#             # smoothing feature angle (in degrees) below which the vertex is not moved,
-#             # only for the Laplacian algorithm, between 0 and 180 degree [default=0]
-#             # laplacian_angle=0,
-#             # smoothing restoring force for the Simple Spring and Polygon Spring
-#             # algorithm, between 0 and 1 [default=0.2]
-#             # polygonspring_force=0.2
-#         )):
-#     """Create a mesh from a bucket
-
-#     Args:
-#         bucket (np.ndarray or aims.Volume): sequence of 3D points.
-#         gaussian_blur_FWWM (numeric) : the gaussian blur filter's full with at half maximum.
-#         threshold_absolute all voxels lower than this value will be set to zero.
-#         threshold_quantile (float in [0,1]): voxels with intensity below this quantile will be set to zero.
-#             Applied only if threshold_absolute is None.
-#         decimation_params (dict, optional): decimation parameters.
-#             reduction_rate : expected % decimation reduction rate
-#             max_clearance : maximum clearance of the decimation
-#             max_error : maximum error distance from the original mesh (mm)
-#             feature_angle : feature angle (degrees), between 0 and 180
-#         smoothing_params (dict, optional): smoothing parameters
-#             type : smoothing alorithm's type (laplacian, simplespring, polygonspring or lowpass)
-#             iterations : smoothing number of iterations
-#             rate : smoothing moving rate at each iteration
-
-#     Returns:
-#         aims Mesh: A mesh obtained from the input volume
-#     """
-#     if not isinstance(bucket, np.ndarray):
-#         bucket = bucket_aims_to_ndarray(bucket)
-
-#     x, y, z = bucket.T
-#     translation = (x.min(), y.min(), z.min())
-
-#     volume = bucket_numpy_to_volume_numpy(bucket)
-
-#     if gaussian_blur_FWWM == 0 and threshold_quantile != 0:
-#         log.warn("Thresholding is automatically disabled with smoothing FWHM=0. To remove this message set threshold_quantile=0")
-#         threshold_quantile = 0
-
-#     return volume_to_mesh_experimental(volume, gaussian_blur_FWWM, threshold_absolute, threshold_quantile, translation, decimation_params, smoothing_params)
 
 
 def get_aims_affine_transform(rotation_matrix, transltion_vector):
@@ -714,40 +525,6 @@ class PyMesh:
         """return a new numpy array converting AIMS mesh properties
         into numpy ndarrays (soma.aims.vector_POINT2DF)"""
         return np.array([x[:] for x in mesh_prop])
-
-
-def transform_datapoints(data_points: np.ndarray, dxyz: np.ndarray, rotation_matrix: np.ndarray, translation_vector: np.ndarray, flip: bool = False) -> np.ndarray:
-    """Transform the data_points.
-
-    The datapoint are scaled according to dxyz, then rotated with rotation_matrix and
-    translated by translation_vector.
-
-    If flip is True, the x coordinates are inverted (x --> -x)
-    """
-
-    tr_data_points = np.empty_like(data_points, dtype=float)
-
-    for i, point in enumerate(data_points):
-        # multiply by scale factors
-        point = point*dxyz
-
-        if (rotation_matrix is not None) and not np.array_equal(rotation_matrix, np.eye(3)):
-            # print("rotate", point)
-            point = np.dot(rotation_matrix, point)
-
-        # translate
-        if (translation_vector is not None) and not np.array_equal(translation_vector, np.zeros(3)):
-            # print("translate", point, translation_vector)
-            point = point + translation_vector
-
-        if flip:
-            # print(point)
-            point[0] = -point[0]
-
-        # store
-        tr_data_points[i] = point
-
-    return tr_data_points
 
 
 class PyMeshFrame:
